Move `Actions` under the `renderer` package
