+++ api/src/main/proto/spine/protodata/actions.proto
@@ -1,0 +1,82 @@
+/*
+ * Copyright 2024, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+syntax = "proto3";
+
+package spine.protodata;
+
+import "spine/options.proto";
+
+option (type_url_prefix) = "type.spine.io";
+option java_package = "io.spine.protodata.renderer";
+option java_outer_classname = "ActionsProto";
+option java_multiple_files = true;
+
+import "google/protobuf/any.proto";
+
+// A collection of code generation actions associated with the parameters passed to these actions.
+//
+// Some code generation scenarios may be solved by applying combination of render actions.
+// For example, we want to make messages declared in proto files named `command.proto`
+// implement the `CommandMessage` interface, and messages declared in files named `event.proto`
+// implement the `EventMessage` interface. Implementation will involve the following steps:
+//
+//  1) Create a render action called `ImplementInterface`. The action would accept the name
+//     of the interface as the parameter.
+//  2) Create Protobuf type for settings that would match a file pattern with the action and
+//     the corresponding parameter with the name of the interface.
+//  3) Create a ProtoData plugin with a renderer which would load the settings and
+//     run corresponding actions.
+//
+// This type addresses the need #2 above, for having parameterised rendering actions in settings
+// passed to ProtoData plugins.
+//
+// Code generation actions listed in this collection serve the same programming language and
+// are applied to Protobuf declarations of the same kind. For example, we may have actions
+// serving code generation in Java for certain `MessageType`s.
+//
+// @see io.spine.protodata.renderer.RenderAction
+// @see io.spine.tools.code.Language
+// @see io.spine.protodata.ProtoDeclaration
+//
+message Actions {
+
+    // Maps a fully-qualified name of a render action class to a parameter
+    // passed to the action.
+    //
+    // The types, names of which come as the keys in the map, must serve the same
+    // programming language and the same type of Protobuf declaration.
+    // The instances of actions are created reflectively. If the above rule is not
+    // followed a runtime action will occur during creation of an action.
+    //
+    // The type of the parameter wrapped into `Any` must match the type of
+    // the parameter accepted by the action. Otherwise, a runtime error will occur.
+    //
+    // If the action does not accept a parameter, the map entry value may contain
+    // either a default instance of `Any` or a packed instance of `Empty`.
+    //
+    map<string, google.protobuf.Any> action = 1 [(required) = true];
+}
