+++ api/src/main/proto/spine/protodata/settings.proto
@@ -35,7 +35,6 @@ option java_package = "io.spine.protodata.settings";
 option java_outer_classname = "SettingsProto";
 option java_multiple_files = true;
 
-import "google/protobuf/any.proto";
 import "google/protobuf/empty.proto";
 import "google/protobuf/descriptor.proto";
 import "spine/protodata/file.proto";
@@ -106,48 +105,4 @@ enum Format {
 
     // A plain string value.
     PLAIN = 5 [(extension) = "txt"];
-}
-
-// A collection of code generation actions associated with the parameters passed to these actions.
-//
-// Some code generation scenarios may be solved by applying combination of render actions.
-// For example, we want to make messages declared in proto files named `command.proto`
-// implement the `CommandMessage` interface, and messages declared in files named `event.proto`
-// implement the `EventMessage` interface. Implementation will involve the following steps:
-//
-//  1) Create a render action called `ImplementInterface`. The action would accept the name
-//     of the interface as the parameter.
-//  2) Create Protobuf type for settings that would match a file pattern with the action and
-//     the corresponding parameter with the name of the interface.
-//  3) Create a ProtoData plugin with a renderer which would load the settings and
-//     run corresponding actions.
-//
-// This type addresses the need #2 above, for having parameterised rendering actions in settings
-// passed to ProtoData plugins.
-//
-// Code generation actions listed in this collection serve the same programming language and
-// are applied to Protobuf declarations of the same kind. For example, we may have actions
-// serving code generation in Java for certain `MessageType`s.
-//
-// @see io.spine.protodata.renderer.RenderAction
-// @see io.spine.tools.code.Language
-// @see io.spine.protodata.ProtoDeclaration
-//
-message Actions {
-
-    // Maps a fully-qualified name of a render action class to a parameter
-    // passed to the action.
-    //
-    // The types, names of which come as the keys in the map, must serve the same
-    // programming language and the same type of Protobuf declaration.
-    // The instances of actions are created reflectively. If the above rule is not
-    // followed a runtime action will occur during creation of an action.
-    //
-    // The type of the parameter wrapped into `Any` must match the type of
-    // the parameter accepted by the action. Otherwise, a runtime error will occur.
-    //
-    // If the action does not accept a parameter, the map entry value may contain
-    // either a default instance of `Any` or a packed instance of `Empty`.
-    //
-    map<string, google.protobuf.Any> action = 1 [(required) = true];
 }
